backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.4.6
httpx==0.27.0
python-multipart==0.0.9
anthropic==0.28.0
pydantic==2.7.1
python-dateutil==2.9.0
